University software implementation project.
An Android application that utilises Google Maps API.
Designed as a social netowkring style blog application for users who wish to attend an event, and would like to do so in
the company of some new friends :).
To assist with the ever difficult task of finding one another, users involved in a specific conversation feed can share their 
location in real time with each other. 

This has been completed as partial fulfilment of my MSc course.